chore: move parseDateTime to core-date and use uuid to validate
