+++ .changeset/red-candles-build.md
@@ -1,0 +1,5 @@
+---
+'@journeyapps/core-date': minor
+---
+
+Add parseDateTime util function
