--- journeyapps-db/src/utils/uuid.ts
@@ -1,3 +1,0 @@
-export function isValid(id: string) {
-  return /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/.test(id);
-}
