+++ journeyapps-db/test/unit/DatabaseAdapterSpec.ts
@@ -2,7 +2,7 @@ import { ObjectType } from '@journeyapps/parser-schema';
 import { hasWebSQL } from './databaseSetup';
 import { WebSQLAdapter } from '../../dist/index';
 import { Query } from '../../dist/query/Query';
-import * as uuid from 'uuid/v1';
+import * as uuid from 'uuid';
 
 if (hasWebSQL()) {
   describe('WebSQLAdapter', WebSQLAdapterSpecs);
@@ -171,7 +171,7 @@ function WebSQLAdapterSyncSpecs() {
   });
 
   it('should perform crud', async function () {
-    const id = uuid();
+    const id = uuid.v1();
     const asset1 = {
       type: 'asset',
       id: id,
